requirements.txt
APScheduler==3.6.3
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
decorator==4.4.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pycparser==2.20
pytesseract==0.3.6
pytz==2020.4
requests==2.24.0
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.25.11
Werkzeug==1.0.1
gunicorn==20.0.4
airtable-python-wrapper==0.15.1
